clean up javascript samples
