+++ samples/javascript_nodejs/46.teams-auth/.eslintrc.js
