+++ samples/javascript_nodejs/46.teams-auth/deploymentTemplates/template-with-new-rg.json
