+++ samples/javascript_nodejs/46.teams-auth/dialogs/mainDialog.js
