+++ samples/javascript_nodejs/46.teams-auth/index.js
